requirements.txt
joblib==1.1.1
numpy==1.23.5
pandas==1.5.3
PyYAML==6.0
PyYAML==6.0.1
scikit_learn==1.2.2
torch==2.0.1
